feat dataset: migrates test project to the dataset folder
